pyproject.toml
[project]
name = "hecate-sentinel"
version = "0.1.0"
description = "Authentication and Authorization API"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "aiosmtplib>=3.0.0",
    "alembic>=1.17.2",
    "argon2-cffi>=23.1.0",
    "asyncpg>=0.31.0",
    "authlib>=1.4.0",
    "email-validator>=2.3.0",
    "fastapi>=0.125.0",
    "fastapi-reqid>=0.1.0",
    "httpx>=0.28.0",
    "psycopg2-binary>=2.9.11",
    "pydantic-settings>=2.12.0",
    "pyotp>=2.9.0",
    "python-jose[cryptography]>=3.3.0",
    "qrcode[pil]>=8.0",
    "sqlalchemy>=2.0.45",
    "user-agents>=2.2.0",
    "uvicorn[standard]>=0.38.0",
]

[dependency-groups]
dev = [
    "aiosqlite>=0.22.0",
    "httpx>=0.28.1",
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=7.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]
